Fix failure to gray out planets at game over.
When the stencil was set on all sprites including the GUI, whatever was in the frame buffer before stayed on the stenciled-out lines, which in most cases was the planets as they appeared before stenciling, rendering the stenciling ineffective. Probably the intention was that the dark background would show through, so we still need to draw that background in full.

Also, a separate check for nil is unnecessary since nil is falsy by itself.